Refactoring composables and packages of FlightQueryScreen and FlightResultsScreen
