+++ booking/presentation/src/main/java/com/projectlab/travelin_android/flight/components/atomos/flightqueryscreen/DateRangePickerField.kt
@@ -1,57 +1,56 @@
-package com.projectlab.travelin_android.flight.components.atomos
+package com.projectlab.travelin_android.flight.components.atomos.flightqueryscreen
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowDropDown
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material.icons.filled.DateRange
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.OutlinedTextFieldDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import com.projectlab.core.presentation.designsystem.theme.customShapes
 
 @Composable
-fun LabelledDropdown(
+fun DateRangePickerField(
     label: String,
-    selected: String,
+    dateRange: Pair<String, String>?,     // e.g. ("11 Apr.", "15 Apr.")
     placeholder: String,
-    options: List<String>,
-    onSelect: (String) -> Unit,
+    onDateRangeSelected: (Pair<String, String>) -> Unit,
     modifier: Modifier = Modifier
 ) {
+    val displayText = dateRange
+        ?.let { "${it.first} – ${it.second}" }
+        ?: placeholder
+
     Column(modifier) {
         Text(
             text = label,
-            style = MaterialTheme.typography.labelSmall
+            style = MaterialTheme.typography.labelSmall  // Poppins 10sp
         )
-        var expanded by remember { mutableStateOf(false) }
         OutlinedTextField(
-            value = if (selected.isNotEmpty()) selected else placeholder,
+            value = displayText,
             onValueChange = { },
             readOnly = true,
             modifier = Modifier
                 .fillMaxWidth()
-                .clickable { expanded = true },
-            trailingIcon = {
+                .clickable {
+                    // aquí dispararías tu DateRangePicker
+                    // por ejemplo: showDatePicker { from, to -> onDateRangeSelected(from to to) }
+                },
+            leadingIcon = {
                 Icon(
-                    imageVector = Icons.Default.ArrowDropDown,
+                    imageVector = Icons.Default.DateRange,
                     contentDescription = null
                 )
             },
             placeholder = {
                 Text(
                     text = placeholder,
-                    style = MaterialTheme.typography.bodyMedium
+                    style = MaterialTheme.typography.bodyMedium // Poppins 12sp
                 )
             },
             shape = MaterialTheme.customShapes.large,
@@ -63,20 +62,5 @@ fun LabelledDropdown(
                 unfocusedLabelColor  = MaterialTheme.colorScheme.onSurfaceVariant
             )
         )
-        DropdownMenu(
-            expanded = expanded,
-            onDismissRequest = { expanded = false },
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            options.forEach { option ->
-                DropdownMenuItem(
-                    text = { Text(option) },
-                    onClick = {
-                        onSelect(option)
-                        expanded = false
-                    }
-                )
-            }
-        }
     }
 }
