+++ booking/presentation/src/main/java/com/projectlab/travelin_android/flight/components/atomos/flightqueryscreen/PassengerPickerBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.projectlab.travelin_android.flight.components.atomos
+package com.projectlab.travelin_android.flight.components.atomos.flightqueryscreen
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
