+++ booking/presentation/src/main/java/com/projectlab/travelin_android/flight/screens/FlightQueryScreen.kt
@@ -1,53 +1,31 @@
 package com.projectlab.travelin_android.flight.screens
 
-import android.app.DatePickerDialog
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Remove
-import androidx.compose.material3.Button
-import androidx.compose.material3.Checkbox
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.projectlab.booking.presentation.flight.FakeFlightViewModel
 import com.projectlab.booking.presentation.flight.IFlightViewModel
 import com.projectlab.core.presentation.designsystem.theme.TravelinTheme
-import com.projectlab.travelin_android.flight.CityInputField
 import com.projectlab.travelin_android.flight.FlightViewModel
 //import com.projectlab.travelin_android.flight.UiTravelClass
-import com.projectlab.travelin_android.flight.components.ClassDropdown
-import com.projectlab.travelin_android.flight.components.atomos.BackButton
-import com.projectlab.travelin_android.flight.components.atomos.ScreenTitle
-import com.projectlab.travelin_android.flight.components.moleculas.FlightQueryBottomBar
-import com.projectlab.travelin_android.flight.components.moleculas.FlightQueryForm
-import java.util.Calendar
+import com.projectlab.travelin_android.flight.components.atomos.flightqueryscreen.BackButton
+import com.projectlab.travelin_android.flight.components.atomos.flightqueryscreen.ScreenTitle
+import com.projectlab.travelin_android.flight.components.moleculas.flightqueryscreen.FlightQueryBottomBar
+import com.projectlab.travelin_android.flight.components.moleculas.flightqueryscreen.FlightQueryForm
 
 /*
 @Composable
